requirements.txt
# Core dependencies
anthropic>=0.20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.8.0
pydantic>=2.0.0
websockets>=10.0,<11.0

# Browser automation
playwright>=1.40.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# MCP dependencies (simplified for initial implementation)
# mcp>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Legacy dependencies (for backward compatibility)
pyautogui>=0.9.54